requirements.txt
Django
python-nmap
psycopg
psycopg2-binary